feat(test): parameterize owner in test cases
This change makes the test cases more robust by parameterizing the owner of the repositories.

The `test/e2e/testcase.go` file has been updated to use a placeholder for the owner in the `tako.yml` configurations for the test cases. The `Setup` and `SetupLocal` functions have been updated to handle this parameterization.

The `cmd/takotest/main.go` file has been updated to add a new `--owner` flag to the `setup` and `cleanup` commands. The `setup` and `cleanup` commands' logic has been updated to use the value of the `--owner` flag when creating the test repositories.