feat(test): add Dirty flag to skip setup/cleanup (#66)
This change introduces a `Dirty` flag to the E2E test cases to allow
skipping the setup and cleanup of test repositories, which can speed up
test runs when repositories are not expected to be modified.

- The `TestCase` struct in `test/e2e/testcase.go` now has a `Dirty`
boolean flag.
- The `Setup` and `SetupLocal` functions now check for the `Dirty` flag
and skip setup if the flag is `false` and the test
repositories/directories already exist.
- The `Cleanup` function now checks for the `Dirty` flag and skips
cleanup if the flag is `false`.
- The E2E tests in `e2e_test.go` have been updated to run sequentially
to prevent race conditions when creating and deleting repositories.
- All test cases have been set to `Dirty: false` to take advantage of
this new feature.

### How to Test

Run the E2E tests in both local and remote mode:

```bash
go test -v -tags=e2e --local ./...
go test -v -tags=e2e --remote ./...
```

The tests should pass, and subsequent runs should be faster.

Fixes: #59