feat: add --repo flag and refactor E2E tests (#71)
This change introduces a `--repo` flag to allow commands to be run from
a remote repository entrypoint. It also includes a significant
refactoring of the E2E test suite to improve reliability and align with
the manual testing workflow.

This change addresses issue #58 by implementing the following:
- A `--repo` flag was added to the `graph` and `validate` commands.
- The flag supports the `owner/repo:ref` format, and the specified ref
is checked out in the cached repository.
- The `takotest` CLI has been updated to support all four E2E testing
modes.
- The E2E test suite now uses `takotest` for setup and cleanup.

### How to Test

1.  Check out this branch: `git checkout issue/58-v2`
2.  Run the local E2E tests: `go test -v -tags=e2e --local ./...`
3.  Run the remote E2E tests: `go test -v -tags=e2e --remote ./...`

Fixes: #58