feat(config): use maps for artifacts and workflows (#78)
This change modifies the `tako.yml` schema to use maps instead of slices
for `artifacts` and `workflows`. This is a more intuitive and
user-friendly approach.

The `Name` field in the `Artifact` and `Workflow` structs is now
populated from the map key during the unmarshalling process and is not
persisted in the YAML file.

Fixes #76